Upgraded to the latest version of PlotVuer.
Indeed, with our composite simulation dataset, the top plot's axis labels are those for the bottom plot and it's only once we have run a simulation that the top plot's axis labels show the correct information. That's because PlotVuer's layout-input did not have reactivity, meaning that redraw/refresh was the only way to update.